book_kdj_rsi_wr_module/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pymupdf>=1.27.0
pytesseract>=0.3.13
Pillow>=10.0.0
